prismlauncher: use jdk17 instead of jdk
Prism Launcher only officially supports jdk 17/18 for running newer
versions of Minecraft, as they found that jdk 19 introduced a number of
edge cases.
This fixes an issue where the newer version of Java that the launcher
proposes is incompatible with the launcher itself, leading to errors
when running instances on newer versions of Minecraft.